requirements.txt
django_filter==21.1
Django==3.2.8
dj_database_url==0.5.0
django_heroku==0.3.1
whitenoise==5.2.0
gunicorn==20.1.0
requests==2.22.0
Pillow==8.4.0